Split unit tests of uploader into two files
