+++ glean/package.json
@@ -95,6 +95,7 @@
     "madge": "^4.0.2",
     "mocha": "^9.0.0",
     "npm-run-all": "^4.1.5",
+    "request": "^2.88.2",
     "request-promise-native": "^1.0.9",
     "selenium-webdriver": "^4.0.0-alpha.8",
     "sinon": "^11.0.0",
@@ -108,7 +109,6 @@
   },
   "dependencies": {
     "jose": "^3.7.0",
-    "request": "^2.88.2",
     "uuid": "^8.3.2"
   },
   "engines": {
