+++ glean/src/platform/qt/uploader.ts
@@ -2,9 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type { UploadResult } from "../../core/upload/uploader.js";
+import log, { LoggingLevel } from "../../core/log.js";
 import type Uploader from "../../core/upload/uploader.js";
+import type { UploadResult } from "../../core/upload/uploader.js";
 import { DEFAULT_UPLOAD_TIMEOUT_MS, UploadResultStatus } from "../../core/upload/uploader.js";
+
+const LOG_TAG = "platform.qt.Uploader";
 
 class QtUploader implements Uploader {
   async post(url: string, body: string, headers: Record<string, string> = {}): Promise<UploadResult> {
@@ -18,21 +21,21 @@ class QtUploader implements Uploader {
       }
 
       xhr.ontimeout = function(e) {
-        console.error("Timeout while attempting to upload ping.", e);
+        log(LOG_TAG, ["Timeout while attempting to upload ping.\n", e.type], LoggingLevel.Error);
         resolve({
           result: UploadResultStatus.RecoverableFailure,
         });
       };
 
       xhr.onerror = function(e) {
-        console.error("Network while attempting to upload ping.", e);
+        log(LOG_TAG, ["Network rror while attempting to upload ping.\n", e.type], LoggingLevel.Error);
         resolve({
           result: UploadResultStatus.RecoverableFailure,
         });
       };
 
       xhr.onabort = function (e) {
-        console.error("The attempt to upload ping is aborted.", e);
+        log(LOG_TAG, ["The attempt to upload ping is aborted.\n", e.type], LoggingLevel.Error);
         resolve({
           result: UploadResultStatus.RecoverableFailure,
         });
